MXS-3394 Make the native match an exact match
This forces the writer of the native template to consider
the sql as a whole. In the simple case it just means that
one has to add a placeholder in the beginning and end of
the match template. This can later be automated away, by
auto adding the placeholders. The point is still that
regex match rather than regex search is used.

Also noticed that a "(.*)" does not always act greedily,
it has to be "(.*?)". Note: in case the Native grammar
needs to change: Any of ECMAScript, EPosix or EGrep
can be used (thus far).

For "scan" and replace to, e.g.  change a table name, a regular
regex can be used (future commit adds RegexRewriter).