MXS-3394 Relax ignore_whitespace regex
No need to police whitespacing. When ignore_whitespace is set,
instead of [[:space:]]+, insert [[:space:]]*. It is important to
note that writing "id=42" or "id = 42" in the match template are
different. With the former "id = 42" will not match even when
ignore_whitespace set, while the later will match both "id=42"
and "id = 42".