Fix improperly showing confirm delete session dialog
If the first session tab is blank but the session being deleted has
unsaved changes in the editor, when clicking the delete session button,
it fails to show the confirm delete session dialog. This is because
when clicking the button it actually sets the session being deleted
as the active session. Deleting inactive session should just delete
it without changing the active session.

This adds `.stop.prevent` modifiers to the @click event so that it
won't trigger the change of active session because the button
is inside the navigation button. Apart from that, the `handleSyncSession`
action needs to be called only when the session being delete is the
active session, otherwise, the current active session will lose
its states.