Update `handleDeleteWke` action
The action is obsolete because the root cause was from the
`resetSessionStates` as it calls `handleSyncSession` for inactive
session. It was fixed in the previous commit and the `handleDeleteWke`
action doesn't actually have to call the `resetSessionStates` action.
In fact, it only has to release the memory by calling
`querySession/releaseQueryModulesMem` because the session object will
be deleted anyway, so there's no point to refresh the session.