Bring back the `WORKSHEET` connection binding_type with supplement connection attributes
Without the `WORKSHEET` type, when having multiple session tabs,
delete the first tab will break the `getWkeFirstSessConnByWkeId`
getter. This getter is used as the connection to be cloned when
adding a new session. However, when the first session tab is deleted,
then adding a new session, it won't use the actual "first" session
tab but the second. This makes it quite hassle to handle delete
all cloned connections properly.

This brings back the `WORKSHEET` type and add `wke_id_fk` attribute
to the connection created by the `conn-man-ctr` component and
`session_id_fk` to the cloned connection created when adding a new
session. These keys would help to identify which connection is
bound to which session tab, which connection is bound to which
worksheet.

When the user opens the connection, it actually creates
2 connections, the first one for the worksheet and the second
for the session tab. The connection bound to the worksheet can only
be deleted by deleting the worksheet or delete it in the dropdown
connection. Delete the session won't affect the worksheet connection.

By using this design, once the worksheet connection is expired,
all of its cloned connections (bound to session tabs) even aren't
expired, will be disconnected