MXS-4187 Fix test_worker
The initiation of the moving must not be done from within the dcall.
Namely, if the message sent to the other worker is delivered and
processed before the dcall has returned, the book-keeping will be
messed up and assertions will fire.

The test also made tighter, shorter delay between dcalls and more
dcalls.