MXS-4235 Replace the usage of `::v-deep` selector` in Query Editor
MaxGUI uses mocha as the test runner. But Vue 2.7 doesn't work well
with `@vue/cli-plugin-unit-mocha` at the moment so all unit tests
will be failed. So far, the only incompatibility issue of the query editor
between Vue 2.7 and Vue 2.6 is the usage of `::v-deep` selector.
https://github.com/vuejs/core/issues/4745#issuecomment-937370398

So this replaces the usage of `::v-deep` selector` in query editor with
vanilla CSS.