MXS-4236 Remove the usage of routes for query editor
The query editor depends on route configuration to work correctly.
e.g. `/query/servers/server_0` is the route for the SQL connection to `server_0`.
The route was introduced in 22.08.0 in order to open the query editor via a
quick link on the cluster detail control page.
However, this is over necessary, there is another way to do a similar job.
When clicking the button to open the query editor for a specific server,
it can just find the worksheet having a connection to that server and
make it an active worksheet.

This gets rid of the need for using routes for the query editor. It persists
the `active_wke_id` state so even after refreshs the page, it won't
lose its active worksheet id.