22.08.0 Tune MaxScale version macros
Leading zeroes (22.08) cause some issues in the release build
process. BB digs out the expected installation minor version
from the VERSION*.cmake file, so it must be what MaxScale reports.
Due to the leading 0 in MAXSCALE_VERSION_MINOR that macro can then
not be used directly in version.h.in, but a separate macro is needed.

MAXSCALE_VERSION_NUMERIC seems to only be used by dead code in
CMakeLists.cmake, so it is removed.