MXS-4219 Use bootstrap parameters when creating vol. servers
When a volatile server is created, the parameter values
'max_routing_connections', 'persistemaxtime', 'persistpoolmax'
and 'proxy_protocol' of the bootstrap servers are used.

At startup it is checked that the values of those parameters
are the same on all bootstrap servers. At volatile server creation
time, the same check is made, but only a warning is logged and the
previous known consistent set of values is used.

Temporary inconsistency among the bootstrap servers must be allowed,
as otherwise it would not be possible to change any setting, as that
has to be done one server at a time.