MXS-4234 Clean up app styles (part 1)
Avoid overriding vuetify component style by using the existing class name
of that component. When the query editor is used with SkySQL UI,
this would cause style conflict. Using a custom name would be a good enough
to avoid this conflict unless SkySQL UI also defines the same custom name
for a component.