MXS-4242 Make preparations for creating initial NoSQL user
To simplify the bootstrapping of a secure NoSQL installation,
the initial NoSQL user will now optionally be created at startup.

If 'authentication_required' and 'authorization_enabled' are
true, then at startup it will be checked whether any NoSQL
user exists. If not, then if 'user' (that is used for connecting
to the backend) has been specified, a corresponding NoSQL user
will be created.

However, that cannot immediately be done, because right at startup
the service master will not be known (the monitor is only starting
up), so using dcalls we wait until it is. When it is available,
we check the grants for the user so that the NoSQL user to be
created can be provided with the correct NoSQL roles.