MXS-4239 Assume slave connection as broken on access denied error
If replication is misconfigured, the slave will keep attempting to
connect to master even if access is denied. This means
Slave_IO_Running will stay in "Connecting"-state and MaxScale will
set it as [Slave] when using default monitor settings.

With this patch, the Last_IO_Errno-field is checked and if the
value is 1045 or 1698, the slave connection is assumed broken.