MXS-4234 Registering vuex module and its plugins automatically
When the query editor is used as an npm package, it needs to be
registered under vue plugin with an option argument that takes
vuex store object and Vuetify object to register query editor
states modules and other necessary modules.

When it's used directly in maxgui, it will skip the registeration
of some modules and components because those are already existed in
maxgui. Besides, maxgui can directly access any modules in the
query editor it needs without using the `maxscale-query-editor`
npm package. However, maxgui still registers it as a plugin to
register query editor states modules and store plugins.