MXS-4234 Clean up query editor
This uses `loading-transparent-overlay` component in query editor
instead of `overlay` component as the query editor only uses
`OVERLAY_TRANSPARENT_LOADING` overlay type, so it's not necessary to
bundle other overlay components.

This also renames the index file of the query editor and move it
to `src` dir. The new name is `maxscale-query-editor.esm.js`.