MXS-4234 Add `mxs_` prefix to custom icons
Custom icons will be exported to the bundle of query editor. To avoid
name conflict with the app using the query editor, this adds
`mxs_` prefix to each icon.

This also renames icon components to following camelCase naming convention.
That eliminates the need to use lodash camelCase function.

The next commit will copy this icon dir to the bundle of query editor.