MXS-4234 Fix app bottom padding
App bottom is ignored since the commit in MXS-4234 switched to show
scrollbar in `v-main__wrap` class. This was done to make sure the
vertical scrollbar will be displayed under app header and above
the footer. However, when `v-main__wrap` set the overflow attribute
to auto, its children elements having `height:100%` will ingore
the parent padding.

This fixes it by set the padding on the wrapper element of
`page-wrapper` component. If the page needs 100% height, it simply
adds `fill-height` class to the component.