MXS-4234 Scoping vue-i18n
The JSON locale file will be included in the query editor package
and merged into the application that consumes the package. To avoid
existing keys being overwritten, it needs to be scoped with a key.
e.g: $t('newConnection') is now $t('mxs.newConnection'). However, this
takes an enormous time to go through each file and change it. Besides,
it's also quite wordy to add mxs. every time a new string is added.

This adds scopingI18n.js to automatically adds the prefix mxs. to the
key argument passing to vue-i18n instance methods. However, the $t
method can not be redefined and it shouldn't be, otherwise, the app
using the query editor would also be affected. So it creates
alternative methods with a prefix.
e.g. $t('newConnection') is now $mxs_t('newConnection').
Overall, replacing all $t, $tc with $mxs_t, $mxs_tc is a lot easier than
go through each argument and add the suffix.