MXS-4234 Add `query-editor-top` slot
SkySQL UI needs this because the `wke-nav-ctr` won't be rendered as part of
their design. However, they still use the `min-max-btn` for setting the
query editor in fullscreen mode. As the result, once the query editor
is maximized, it doesn't include the `min-max-btn` as it's rendered
outside the `query-editor` component.

This adds `query-editor-top` to allow render query editor
sub-components that need to be shown when it's in fullscreen mode.