MXS-4260: Reset causal read state on trx replay
There were two cases where the causal read state wasn't reset correctly
when a server went down: when the server died during a
`START TRANSACTION READ ONLY` statement or during a causal
read done inside of an optimistic transaction being executed on a slave
server. If the transaction ended up being replayed on the master, the
first packet of the response would end up being discarded because it was
mistaken for the reply to the MASTER_GTID_WAIT query which results in a
hanging connection.

Additionally, it was possible for a transaction replay to start if a
Galera node reported a WSREP error while a transaction was open. This
would've caused a transaction replay to be started without the causal
read state being reset.