MXS-4234 Configurable query editor base url
When the query editor is used in the unified portal, there
will be multiple maxscale api addresses. So the base url for
the query editor should be dynamic.

This stores base url in `queryEditorConfig` state module,
the value can be used to change the base url by overriding the config
in the request interceptor of `queryHttp` axios instance.