MXS-4234 Add fallback background color for query editor tooltips
MaxGui customizes vuetify sass variables for `v-tooltip` component
to automatically set the background of the tooltip to white. However,
when the query editor is bundled as a package, it won't include
these customized variables to avoid style conflicts with the app
using the query editor.

This uses vuetify helper class to set the background.
This also removes the `$tooltip-padding` variable as it's not
neccessary.