MXS-4234 Add query editor dependencies
This defines query editor's dependencies so that the required
packages can be installed automatically if they don't already
exist in the `node_modules` dir of the application using the
query editor.

`peerDependencies` ensures the compatibility of the required
packages.

This also renames the lib to `mxs-query-editor` and updates
external dependencies.