MXS-4234 Rollup cjs format and optimize the bundle
cjs format is required for test env in the unified portal.

This optimizes the build by minifying css output and injecting
the copyright content to the output files.

This also adds `rollup-plugin-terser` package to minify the output
and removes `rollup-plugin-postcss` package because maxgui doesn't
use it. Using `rollup-plugin-scss` package is enough.