Make mxs1549_optimistic_trx more robust
Due to the way that readwritesplit handles reads after session commands,
the test must wait for a while after starting the transaction with
`SET AUTOCOMMIT=0` to guarantee that both the master and the slave
servers have executed it. Otherwise the read-only transaction is routed
to the master server.

Also formatted the test with the new Uncrustify configuration (mostly
lambda alignment changes).