MXS-4234 Add `meta` parameter for `openConnect` action
This `meta` object can be used to store extra information about
the connection being created. When creating a connection in SkySQL,
it needs to find out the maxscale service object name in order to
create a connection. However, in multi-tenant scenarios, the maxscale
service object name may be identical, so it's not possible to know
which connection belongs to which SkySQL service (topology).

The SkySQL service name can be stored in the meta object.