MXS-4275: Ignore cluster config with --export-config
As the config export uses the INI files to produce the output, it cannot
currently support a cluster configuration as a source. For this reason
the --export-config will now always behave as if no cached cluster
configuration is present.

Documented the limitation and added a link to the issue that describes
it.