MXS-4279: Use "sub" field in JWTs for the username
This is more in line with the intended purpose of the claim. The "aud"
value is still set to the same value as the "sub" field to retain
backwards compatibility with external applications but it is no longer
used as the source of the username. This opens up future use of the
"aud" field for other, more appropriate purposes.