Replace debug assertions in mxs2729_rebalance
A failing debug assertion crashes the test and prevents the post-test
cleanup from being done. This makes it harder to figure out why the
tests fail. Using the normal `expect` function is much better as it both
displays the values and allows the program to continue.

Also formatted the test with Uncrustify.