MXS-4278 Fix connection name for disappearing in an edge case
When creating a new connection for the worksheet that already has been
bound with a connection, it unbind the current connection from the
worksheet even the new connection has not been created successfully.

The unbind method should be called only if the new connection is
created successfully.

This also moves `unbindConn` and `bindConn` methods to `queryConn`
state module so that the unified portal UI can reuse these methods.