MXS-4281 Fix query editor sql connections
Connections aren't bound to the query tabs properly in an edge case.
It happens when there are two worksheet connections and 1 query tab.
If the user adds a new query tab then switches back to the other worksheet
connection, it failed to bind the existing connection for the second
query tab.

This adjusts `onChangeConn` action to handle that case.