MXS-4283: Use lcall in kill_complete
This makes sure the callback is called immediately after all epoll
events have been processed. This removes the need to increment the
reference count of the session as it will be semi-alive when lcalls are
processed. With a worker message it was possible that the session was
freed before the queued event was read.