MXS-4289: Always use master for trx sescmd replies
This prevents the case where a transaction that would otherwise start on
the master server would end up starting on a slave server due to there
only being slave servers available at that time. This keeps the logic
consistent and makes error handling simpler if read-write transaction
are never started on slave servers, even if technically possible to
handle in a semi-graceful manner.