MXS-4290 Explicitly specify expression depth limit for qc_sqlite
The expression depth limit puts a cap on e.g. how may comparisons
there can be in a WHERE clause. That puts a cap on how deep the
recursion can be and hence also on the amount of memory needed
for the stack. If set to 0 there is no limit. However, tests
revealed that with no limit, a statement like "SELECT 1 FROM t
WHERE 1=1 OR 1=1 OR ... OR 1=1" with some 30000 comparisons,
will cause the stack to overflow on a typical Linux machine,
while 25000 comparisons will not. Hence, the limit is set at
25000, which is far more than what is needed by the statement
in this issue.