MXS-4293: Fix causal_reads=universal memory leak
The buffer passed to the function was unused and thus leaked memory when
the return value of the function was assigned to it.

This also revealed a problem in mxs::Reply::row_data(): if something
caused the resultset to be freed or replaced, the string_views that
pointed to that buffer would now point to a freed GWBUF. To handle this
case, a shallow clone of the GWBUF is kept around for the duration of
the clientReply call if any rows were collected.