MXS-4044: Log admin related state changes
Printing the message when the request is being processed is much easier
than trying to fit it into the normal state change processing logic
which ignores both of the bits.

The draining state isn't printed, or for that matter used, when the
maintenance mode is set. For this reason the state changes for drained
servers are only printed if the server is not in maintenance mode.