MXS-4161 Add attributes.system to maxscale resource
So, now there will be an object like in the "data.attributes"
object returned by the "maxscale" resource.

            "system": {
                "machine": {
                    "cores_available": 8,
                    "cores_physical": 8,
                    "cores_virtual": 8.0,
                    "memory_available": 41717088256,
                    "memory_physical": 41717088256
                },
                "maxscale": {
                    "query_classifier_cache_size": 6257563238,
                    "threads": 8
                },
                "os": {
                    "machine": "x86_64",
                    "nodename": "johan-P53s",
                    "release": "5.4.0-125-generic",
                    "sysname": "Linux",
                    "version": "#141~18.04.1-Ubuntu SMP Thu Aug 11 20:15:56 UTC 2022"
                }
            },

The point is to return information using which it will be
possible to quickly see in one place whether something might
be seriously wrong. E.g. if maxscale.threads is significantly
larger that properties.virtual_cores, MaxScale is probably being
run with "threads=auto" in a container whose resources have
been restricted.