MXS-4267 Replace `delimiter` inputs with `Fields terminated by` input
The old UI has a delimiter dropdown with 3 options which are
"Tab, Comma and Custom`. If the user chooses `Custom`, it displays
another text input for entering custom delimiter. This text input
is escaped automatically by Vue.js, so if the input is `\r`, it's
actually is `\\r`. So JS won't recognize it as a carriage return
character.

This adds `unescapedUserInput` method to unescape user input. With
this method, there is no need to have 2 inputs. The two mentioned inputs
are now replaced with just an input named `Fields terminated by`.