MXS-4303 Fix the disconnect connection action in the reconnection dialog
When the connection is lost during the query, the reconnection dialog
is opened to allow the connection to be reopened or disconnected.
Clicking on the `Disconnect` button, the connection abound to the
query tab is disconnected but the connection bound to the worksheet
isn't.

This disconnects the worksheet connection whenever the button is clicked.