MXS-4290 Reduce sqlite max expression depth to 4096
The default max expression depth is 1000 and for the issue in
question 1200 would have been enough. It was originally changed
to 25000, which is fine in a non-constrained environment, but
may be too much in a constrained container environment. 4096
is a compromise; large enough to be sufficient in practice but
not so large that some pathological case can cause MaxScale to
crash in a constrained environment.