MXS-4313: Convert names to strings
The names of objects in the REST API are always strings. If something
that looks like a number was given to MaxCtrl, it was automatically
converted into a number. Explicitly converting them into strings solves
this problem.

Added a test case that verifies that creation of users with numbers as
names works.