MXS-4251 Kill the query immediately
The `KILL QUERY thread_ID` statement responses immediately but that
doesn't mean the query is stopped. The querying request from the API
is in a pending state until it receives data from the database saying
that the query is now killed. This is an expected result, but sometime
it may take time for the query to be killed, the user can't run any query
during that period in that query tab. That is bad in terms of UX.

This commit uses `AbortController` API to abort the query request. If
the user clicks the stop button, it sends `KILL QUERY thread_ID` then
aborts the running query.