MXS-4271: Read certificates from OIDC server
Instead of providing a manual list of public certificates, they can be
retrieved directly from the OIDC server. This makes reloading them
easier as it removes the need to copy certificates when they can be
retrieved directly from the server. The only downside is that the JWK
format is quite annoying to convert into PEM if the x5c claim is not
present.