MXS-4271: Validate JWTs based on the "kid" field
By using the "kid" field of the token, we know which key we should use
to decode it. This also means that if the key is not known, no attempt
to decode it is done.

As the key algorithm is specified in the JWK, the class that validates
the tokens can be constructed on startup. This also removes the problem
where the "alg" value in the JWT was used and not the one stored in the
JWK.

The "iss" field of the tokens must now also match the value of
`admin_odic_url`. Previously the issuer of the token was not checked at
all.