MXS-4271: Use issuer value from OIDC metadata
The OpenID Connect specification
(https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata)
states that the "issuer" field of the metadata should be used as the
expected value for the "iss" field for all tokens. This also solves any
problems that would occur if the value of admin_oidc_url and the actual
issuer are not identical but are logically the same.