MXS-4299 Use `session` data via props for the `load-sql-ctr`
When the inactive query tab has unsaved changes get deleted, its content
isn't saved properly because the `openConfDlg` in the `session-nav-ctr`
has a `$refs` to the active query tab instead of the inactive query tab.
As a result, the content belongs to the active query tab. Using `$refs`
only work for active query tab because there is no $refs for the inactive
one.

The `load-sql-ctr` component will be later separated into two components.
A component for showing buttons and the other for the dialog. Its methods
will be also made reusable so that the `session-nav-ctr` can also call
the `handleSaveFile` method.

This commit uses `session` data via the props instead of using the
current active session state.