MXS-4299 Fix `openConfDlg` method
All methods related to the `handleSaveFile` method are now moved
to a mixin file called `saveFile`. The `session-nav-ctr` component
can use that mixin to call the method to save the file before deleting the
session. This adds `file_dlg_data` state to the `editor` state module and
creates the `file-dlg-ctr` component to show the dialog.

This also renames the `load-sql-ctr` to `file-btns-ctr` as it
now renders only buttons.