MXS-4333: Make ENGINE configurable in kafkaimporter
When something other than InnoDB is used, the table layout will be
changed to a less specific version to acommodate engines that do not
support all the featuers of the old one. The main difference is the lack
of the generated column which had a unique index defined for it. The
Aria and ColumnStore engines are not compatible with it: Aria supports
generated columns but not indexes on them and ColumnStore does not
support generated columns nor unique indexes.

The JSON type with the automatic CHECK constraint is much more common
now which means we don't need to define the underlying type anymore and
can simply substitute it with `JSON NOT NULL`.

The added test case verifies that the feature works and behaves as
expected when the definition is modified at runtime. Also fixed the
shadowing declaration of `i` in the tests.