MXS-4340 Configure listeners and services
The listeners and services need some actions when a new routing
worker is created. With this change, a new routing worker will
be taken into use so that it will accept new connections. Other
similar configuring is likely to be needed in other contexts
as well; e.g. cache filter may need some tuning.