Make Access-Control-Allow-Origin configurable
This is needed to prevent CSRF attacks when CORS mode is
enabled. Changed the default value of the header to `*` to make it clear
all origins are allowed. The behavior where the request origin was
unconditionally used was confusing. Since this is only of use when
embedding into other webpages and when testing the GUI outside of
MaxScale, this can remain in the debug option list.

Fixed the odd signature of the mxs_admin_enable_cors() function and
rearranged the standard headers to be in alphabetic order.