MXS-3269 Enanble thread increase after decrease
If the number of threads is increased after it has been decreased,
existing threads are taken into use before new are created.

This is an intermediate commit as there now is quite a lot of
message passing between the main worker and an activated routing
thread. It will be changed so that there is but one message from
the main worker to the worker that is activated.