MXS-3269 Rename Worker::state_t to Worker::EventLoop
And rename state() to event_loop_state(). One major reason
is not to monopolize the general purpose "state". event_loop_state()
provides very little information in addition to what started() does;
the only difference is that you can distinguish between the loop
not running becuase the Worker has not been started (which means
that it shortly will be started and hence it is ok to send messages)
and the loop not running because the worker has stopped (and in the
context of MaxScale won't be started again). Would be better to
arrange this some other way so that the duplication could be removed.