MXS-3269 Rename active workers to running
When multiple workers are discarded at the same time, it is
possible that a worker with a smaller index becomes inactive
before a worker with a larger index does. Thus, among workers
with a smaller index that nActive (now nRunning) there may
temporarily be inactive ones. Thus, better to use another word
than nActive.