MXS-3269 Reduce # of running only when it has occurred
When the number of threads has been reduced (via the REST-API),
the actual number of running threads can be reduced only when
the workers have stopped routing requests, which happens only
when all sessions have terminated. A worker will no sessions
can be deactivated immediately, but a worker with sessions
can be deactivated only after the number of sessions has
dropped to 0.

It need to be separately stored what the desired number of
threads is, so that new sessions will not be assigned to workers
that are in the process of deactivating, but that is for a
separate commit.