Log message when blocking node in sts_readwritesplit
By placing the message in the MaxScale log itself, we'll see the timing
of when the node was blocked and when the query arrived. Based on the
log contents the blocking either takes place too late or the monitor
detects it too late to react to it both of which would explain why the
query works when it shouldn't.