MXS-4353: Fix syslog message filtering
The SYSLOG_IDENTIFIER is now explicitly logged by MaxScale using the
given log identifier. This fixes the problem where the chosen identifier
was not used for the messages if it differs from the program's name.

The behavior of mxb_log_init was not exactly what it was documented to
be. The identifier deduction relied on syslog and would not be included
in any other mechanism. Given that none of the other mechanisms log the
identifier, this was not a problem. Now that the SYSLOG_IDENTIFIER field
is explicitly set, the value should be stored by the log subsystem.

Changed MaxScale to pass the log identifier explicitly instead of
relying on the program name. This makes it so that if we, for whatever
reason, decide to change the program name, the log message identifier
will remain the same.

Messages logged by MaxScale 2.5 and older do not contain the TIMESTAMP
field. By filtering out any messages which do not have this field, we
can guarantee that each JSON object will have at least the timestamp,
message and priority fields.