MXS-4354 Prep CMakeLists.txt for clang
Disabling a few warnings that will stay disabled.
Disabling a list of warnings that will be removed one by one.

Removed the warning "maybe-uninitialized" in
qc_sqlite/CMakeLists.txt, replaced with
"uninitialized" as the former is not supported by clang
and seems dubious. If needed these can be chosed based
on the current compiler (CMAKE_C_COMPILER_ID).

The main CMakeLists.txt would benefit from a little refactoring,
mainly to make sure disabled warnings are applied last so that
they are not re-enabled depending on order of various
flag variables.