MXS-4253 Fix unexpected redirection to 404 page after PAM user is logged in
This issue occures because the GUI sends a request to `/users/inet/:name`
right after the user clicks the Login button. So if the logged in user
isn't an inet user, it returns 404.

This switches to use `authHttp` axios instance instead of the `$http`
because it doesn't redirect to 404 page automatically.