MXS-3269 Update how qc cache size is calculated
At runtime the thread specific query classifier size must be
calculated using RoutingWorker::nRunning() and not Config::n_threads.
When decreasing threads, the latter will immediately be updated while
the former will be only when the number of running threads really has
decreased. If the latter were used, then the amount of memory used
for the classifier cache would go up while the threads are draining.