MXS-4285 Disable document pointer events while resizing panels
This will come in useful in the next commit where it leverages the
resizing feature on the schema sidebar to make it similar to vscode
sidebar. i.e. The sidebar will be automatically collapsed when it reaches
certain width. Because the query editor sidebar is placed next to
app sidebar, when resizing, the user may mouse over the app sidebar
and triggers the expansion of the sidebar.
Though this is not relevant to the ticket but it improves UX and it is a
nice to have feature.