Block outbound traffic in block_node
This fixes the problem where a blocked node would end up generating
responses to queries if they were initiated before the blocking took
place.

This causes problems in at least in sts_readwritesplit: altering the
test to use a 20 second monitor interval would always cause the result
of a `SELECT SLEEP(10)` to be returned even if the node is blocked. With
this commit and a 20 second monitor interval, the test passes.