Update REST-API documentation
Updated the documentation and adjusted the configuration to use a fixed
amount of threads.

Also added `users_refresh_time=0s` to avoid the problem where the
servers aren't up when MaxScale has started but are when the user logs
in. The refresh limit appears to be hit in this case.