MXS-4285 Extend `mxs-split-pane` to stop resizing on `maxPercent`
At the moment, the value is allowed to go up to 100%. This causes
the right pane (or bottom pane) to be completely hidden. This also
causes odd view on the UI. e.g. During the resizing of the query editor
sidebar panel from the initial value to 100%, the editor toolbar and
session tabs don't have enough space to render buttons, icons,...

This adds `maxPercent` props as the maximum threshold value that the
left(top) panel can be resized to.