Add missing session ID to log messages
Whenever something was queued for later execution, the log messages
would not contain the current session's ID in them. This is due to the
fact that the current DCB being processed is what determines the current
session and when the callback is executed, the DCB is no longer the
active one. Instead of modifying the currently active DCB, the current
session can be modified which is guaranteed to be valid for longer than
any single DCB might be. This could also be done using the client_dcb of
the session but the active DCB management code is not exposed to the
rest of the MaxScale.

The lack of session IDs in log messages made it harder to form a
coherent view of the whole session when parts of the message did not
identify who they belonged to. The KILL command processing was the main
motivator for this as it is involved in a handful of failing test cases.

This commit also has the added benefit of correctly showing the active
session for any fatal signals generated by code that wasn't generated
inside of the epoll event handling loop.