MXS-4285 Refactor `mxs-split-pane` to add events on window object
This adds `mousemove` and `mouseup` events to the window object
instead of binding them to the container element. This ensures the
events are triggered even when the mouse cursor is moused out the
container element. Another benefit of this is that those events can
now be listened only when the `mousedown` event is triggered, so
it reduces less memory usage. However, this requires providing the
boundary width/height in order to calculate the percentages.

Overall, this brings a smoother resizing effect than the old way.