Stop failing tests immediately
If the normal test startup fails, for example due to a bad
configuration, the test should perform cleanup and exit. This will
prevent tests from crashing when a part of the test system hasn't been
properly initialized.

The configuration used by the config_test test case was not valid as the
only replaced if the test has the REPL_BACKEND label, the test now
requires a valid configuration.