Add tracing for mxs548_short_session_change_user
The test appears to have timed out with no explanation as to why it did
so. Since the test is a stress test of the COM_CHANGE_USER mechanism,
enabling session level tracing should be the least invasive action that
still allows us to see why the test times out.