MXS-3269 Add DESIRED as option when dealing with rworkers
The 'desired' number of workers is the goal, but when the
number of threads is decreased, it may take a while before
it is reached. This is relevant when e.g. listeners are
added, because they should not be added to running workers
that in the process of going down.