Fix SLES 15 builds
The update step of the hiredis library is failing with SLES 15. The
update itself is triggered by the modification of CMake configuration to
build a different package which means it's not a real update.

In addition, we want the package to remain the same across CMake
invocations to avoid the extremely unlikely scenario of someone
modifying the tag to point to something else mid-build.