MXS-4285 Extend `mxs-split-pane` to control the deactivated resize zone
In vscode, when resizing a sidebar to its minimum width, the sidebar stops
resizing at its minimum width but if the user resizes it to the midpoint
of the sidebar, it auto collapses. In other words, the midpoint controls
the collapsion and expansion.

This brings some benefits in terms of user experience.
For example, when the user wants to resize quickly to the minimum or
maximum value without collapsing the pane, the pane won't be collapsed
until it reaches the midpoint. On the other hand, if the user actually
wants it to collapse, this midpoint speeds up the process a bit.

This midpoint was done in the `wke-ctr` but it should have been a
feature of `mxs-split-pane`.