MXS-4358 Fix `authCheck` action
When the auth token is expired but the user is still on the query
editor page, it doesn't redirect to the login page.

The `authCheck` action is finished but the `SET_LOGGED_IN_USER`
mutation isn't called because after that there is an API call
by the query editor to `/sql` endpoint and stops at exception.

This adds interceptor for the `authCheck` action to redirect
the user to the login page.