MXS-4013 Don't fetch `VIEW` for the `Tables` node
Currently, when expanding the `Tables` node, it fetches all type
of tables including `VIEW` type. There will be a dedicated node
for showing `VIEW` type, so there is no need to fetch `VIEW` when
expanding the `Tables` node.

This also renames some actions because the current names don't explain
the actions clearly.