Rename values() to collect_values()
This removes the confusing naming of the values() function in
WorkerGlobal and the ones in the various mxs::Configuration
implementations that use it to return the WorkerGlobal class that
contains the thread-local values. It also clarifies the fact that it is
an active operation and not a completely passive one.