MXS-4317: Delay query execution if KILLs are pending
By tracking whether the KILL commands have finished, the smartrouter can
delay the execution of new queries before the killing of the old ones is
done. This avoids the problem where the wrong query ends up being
killed.

This commit fixes the sr_basics test which already serves as the test
case for this problem.