MXS-4013 Automatically assign level value for each node
The level of node can be automatically assigned if the id of the node
contains id of all parent ids. At the moment, the id attribute is used
as a qualified name, hence it can't be used to count the level.

This adds `hrchy_id` attribute which will be later renamed to `id`
once `id` is renamed to `qualified_name`